Security update.

+++ Documents/GitHub/pz_core/actions/admin/user/makeadmin.php
@@ -6,12 +6,12 @@
  * column in the users_entity table.
  *
  * In <1.7.1, admin is a piece of metadata on the user object.
- *
+ * updated
  * @package Elgg.Core
  * @subpackage Administration.User
  */
 
-$guid = get_input('guid');
+$guid = (int) get_input('guid');
 $user = get_entity($guid);
 
 if (($user instanceof ElggUser) && ($user->canEdit())) {
